Added details.txt; update floorplan labels
details.txt is a rough draft of the equipment specifications list. It lists the equipment as it appears on the floorplan, and includes footprints, weights, descriptions, and other relevant data. It is partially complete.

Changed some names in the floorplan (layout_new_general.SLDDRW) to be more descriptive and remove ambiguity for finding the corresponding item on details.txt